Finalizado as funções do InstaLab-Hooks
