+++ modulo4/use-state/insta-lab-hooks/src/components/Post/Post.js
@@ -4,21 +4,49 @@ import { PostContainer, PostHeader, UserPhoto, PostPhoto, PostFooter, CommentCon
 import IconeComContador from '../IconeComContador/IconeComContador'
 import SecaoComentario from '../SecaoComentario/SecaoComentario'
 
-import iconeCoracaoBranco from '../../img/favorite-white.svg'
 import iconeCoracaoPreto from '../../img/favorite.svg'
+import iconeCoracaoBranco from '../../img/favorite-white.svg'
 import iconeComentario from '../../img/comment_icon.svg'
 
-const Post = (props) => {
-
+function Post (props) {
+  const [curtido, setCurtido] = useState(false)
+  const [numeroCurtidas, setNumeroCurtidas] = useState(0)
+  const [comentando, setComentando] = useState(false)
+  const [numeroComentarios, setNumeroComentarios] = useState(0)
+  const [comentarios, setComentarios] = useState([])
 
   const onClickCurtida = () => {
-  };
+    setCurtido(!curtido)
+    if (curtido) {
+      setNumeroCurtidas(numeroCurtidas - 1)
+    } else {
+      setNumeroCurtidas(numeroCurtidas + 1)
+    }
+  }
 
   const onClickComentario = () => {
-  };
+    setComentando(!comentando)
+  }
 
   const enviarComentario = (comentario) => {
+    setComentarios([...comentarios, comentario])
+    setComentando(false)
+    setNumeroComentarios(numeroComentarios + 1)
   }
+
+  const iconeCurtida = curtido ? (iconeCoracaoPreto) : (iconeCoracaoBranco)
+  
+  const caixaDeComentario = comentando ? (
+    <SecaoComentario enviarComentario={enviarComentario}/>
+  ) : (
+    comentarios.map((comentario) => {
+      return (
+        <CommentContainer>
+          <p>{comentario}</p>
+        </CommentContainer>
+      )
+    })
+  )
 
   return (
     <PostContainer>
@@ -31,18 +59,19 @@ const Post = (props) => {
 
       <PostFooter>
         <IconeComContador
-          // icone={iconeCurtida}
+          icone={iconeCurtida}
           onClickIcone={onClickCurtida}
-          // valorContador={numeroCurtidas}
+          valorContador={numeroCurtidas}
         />
 
         <IconeComContador
           icone={iconeComentario}
           onClickIcone={onClickComentario}
-          // valorContador={numeroComentarios}
+          valorContador={numeroComentarios}
         />
+
       </PostFooter>
-      {/* {caixaDeComentario} */}
+      {caixaDeComentario}
     </PostContainer>
   )
 }
